test(cli): share integration test support
